fix(locale): rename tj to tg
