+++ src/runtime/locale/index.ts
@@ -50,8 +50,8 @@ export { default as sk } from './sk'
 export { default as sl } from './sl'
 export { default as sq } from './sq'
 export { default as sv } from './sv'
+export { default as tg } from './tg'
 export { default as th } from './th'
-export { default as tj } from './tj'
 export { default as tr } from './tr'
 export { default as ug_cn } from './ug_cn'
 export { default as uk } from './uk'
